# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * AdForge creative intelligence API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Brand {
  id: string;
  name: string;
  website: string;
  description: string;
  industry: string;
  audience: string;
  personality: string[];
  visualStyle: string;
  typography: string;
  primaryColor: string;
  secondaryColors: string[];
  accentColor: string;
  /** @nullable */
  logoPath: string | null;
  isDemo: boolean;
  updatedAt: string;
}

export interface BrandInput {
  /** @minLength 1 */
  name: string;
  website: string;
  description: string;
  industry: string;
  audience: string;
  personality: string[];
  visualStyle: string;
  typography: string;
  primaryColor: string;
  secondaryColors: string[];
  accentColor: string;
  /** @nullable */
  logoPath?: string | null;
}

export type BrandUpdate = BrandInput;

export interface BrandAsset {
  id: string;
  brandId: string;
  name: string;
  type: string;
  /** @nullable */
  objectPath: string | null;
  previewUrl: string;
  tags: string[];
  isFavorite: boolean;
  isDemo: boolean;
  createdAt: string;
}

export interface BrandAssetInput {
  /** @minLength 1 */
  name: string;
  type: string;
  /** @nullable */
  objectPath?: string | null;
  previewUrl: string;
  tags?: string[];
}

export interface Campaign {
  id: string;
  brandId: string;
  name: string;
  productName: string;
  description: string;
  productUrl: string;
  benefits: string[];
  /** @nullable */
  price: string | null;
  /** @nullable */
  offer: string | null;
  cta: string;
  audience: string;
  ageRange: string;
  location: string;
  interests: string[];
  painPoints: string[];
  desires: string[];
  objective: string;
  platform: string;
  format: string;
  aspectRatio: string;
  status: string;
  isDemo: boolean;
  createdAt: string;
}

export interface CampaignInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  productName: string;
  description: string;
  productUrl: string;
  benefits: string[];
  /** @nullable */
  price?: string | null;
  /** @nullable */
  offer?: string | null;
  cta: string;
  audience: string;
  ageRange: string;
  location: string;
  interests: string[];
  painPoints: string[];
  desires: string[];
  objective: string;
  platform: string;
  format: string;
  aspectRatio: string;
}

export type CampaignUpdate = CampaignInput;

export interface CreativeConcept {
  id: string;
  campaignId: string;
  name: string;
  family: string;
  angle: string;
  visualDirection: string;
  hook: string;
  headline: string;
  bodyCopy: string;
  cta: string;
  composition: string;
  colorDirection: string;
  emotion: string;
  audienceInsight: string;
  creativeCount: number;
}

export interface Creative {
  id: string;
  campaignId: string;
  conceptId: string;
  conceptName: string;
  family: string;
  hook: string;
  headline: string;
  bodyCopy: string;
  cta: string;
  previewUrl: string;
  platform: string;
  format: string;
  aspectRatio: string;
  emotionalDriver: string;
  creativeAngle: string;
  audienceInsight: string;
  readinessScore: number;
  isFavorite: boolean;
  isDemo: boolean;
  status: string;
  createdAt: string;
}

export interface CreativeGenerationInput {
  /** @minLength 1 */
  campaignId: string;
  conceptIds?: string[];
  /**
     * @minimum 1
     * @maximum 12
     */
  count?: number;
}

export interface CreativeUpdate {
  headline?: string;
  bodyCopy?: string;
  cta?: string;
  isFavorite?: boolean;
}

export interface CreativeVariation {
  id: string;
  creativeId: string;
  dimension: string;
  label: string;
  headline: string;
  bodyCopy: string;
  cta: string;
  previewUrl: string;
  readinessScore: number;
  createdAt: string;
}

export interface Activity {
  id: string;
  label: string;
  detail: string;
  timestamp: string;
  type: string;
}

export interface DashboardSummary {
  brandHealth: number;
  creditsRemaining: number;
  activeCampaigns: number;
  generatedCreatives: number;
  savedCreatives: number;
  recentCampaigns: Campaign[];
  recentCreatives: Creative[];
  activity: Activity[];
}

export interface PerformancePoint {
  label: string;
  value: number;
  /** @nullable */
  secondaryValue: number | null;
}

export interface IntelligenceSummary {
  isSampleData: boolean;
  totalImpressions: number;
  averageCtr: number;
  conversionRate: number;
  winningFamily: string;
  winningHook: string;
  familyPerformance: PerformancePoint[];
  weeklyPerformance: PerformancePoint[];
  recommendations: string[];
}

export interface UploadRequest {
  /** @minLength 1 */
  name: string;
  /**
     * @minimum 1
     * @maximum 10485760
     */
  size: number;
  contentType: string;
}

export interface UploadResponse {
  uploadURL: string;
  objectPath: string;
}

export type ListCreativesParams = {
campaignId?: string;
};