# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AdForge creative intelligence API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [storage]
      summary: Request a presigned upload URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadRequest"
      responses:
        "200":
          description: Upload instructions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResponse"
components:
...
  schemas:
...
    UploadRequest:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        size: { type: integer, minimum: 1, maximum: 10485760 }
        contentType: { type: string }
      required: [name, size, contentType]
    UploadResponse:
      type: object
      properties:
        uploadURL: { type: string }
        objectPath: { type: string }
      required: [uploadURL, objectPath]