# lib/api-zod/src/generated/types/uploadRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * AdForge creative intelligence API
 * OpenAPI spec version: 0.1.0
 */

export interface UploadRequest {
  /** @minLength 1 */
  name: string;
  /**
     * @minimum 1
     * @maximum 10485760
     */
  size: number;
  contentType: string;
}